diri: move Orderbook events parsing in another package
