--- crates/diri/src/mod.rs
@@ -1,1 +1,0 @@
-mod orderbook;
