+++ crates/orderbook/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "orderbook"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+cainome = { git = "https://github.com/cartridge-gg/cainome", tag = "v0.4.0", features = [
+    "abigen-rs",
+] }
+
+starknet = "0.12.0"
+starknet-types-core = "0.1.7"
+
+num-bigint = "0.4.4"
+num-traits = "0.2.19"
